requirements.txt
openai
streamlit
PyPDF2 
python-docx 
pandas
altair<5